+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 - **New:** Seed for thread prefix (#50)
 - **New:** Seed for resource prefix (#51)
 - **New:** Seed for thread prefix group (#52)
+- **New:** Seed for resource prefix group (#53)
 
 ## 1.1.0 Alpha 4 (`1010014`)